some testing code and modification in testcases

+++ meduim/reduction/fs_main.cu
@@ -4,7 +4,7 @@
 #include <hip/hip_runtime.h>
 #include <fstream>
 
-__global__ void reduction(const float* input, float* output, int N){
+__global__ void reduction(const int* input, int* output, int N){
     int idx = blockDim.x * blockIdx.x + threadIdx.x;
 
     if (idx >= N) return;
@@ -12,14 +12,14 @@ __global__ void reduction(const float* input, float* output, int N){
     atomicAdd(output, input[idx]); 
 }
 
-extern "C" void solve(const float* input, float* output, int N) {  
-    float *d_input, *d_output;
+extern "C" void solve(const int* input, int* output, int N) {  
+    int *d_input, *d_output;
 
-    hipMalloc(&d_input, N * sizeof(float));
-    hipMalloc(&d_output, sizeof(float));
+    hipMalloc(&d_input, N * sizeof(int));
+    hipMalloc(&d_output, sizeof(int));
 
-    hipMemcpy(d_input, input, N * sizeof(float), hipMemcpyHostToDevice);
-    hipMemset(d_output, 0, sizeof(float));
+    hipMemcpy(d_input, input, N * sizeof(int), hipMemcpyHostToDevice);
+    hipMemset(d_output, 0, sizeof(int));
     
     int threadperblock = 256;
     int blockpergrid = (N + threadperblock - 1) / threadperblock;
@@ -27,7 +27,7 @@ extern "C" void solve(const float* input, float* output, int N) {
     reduction<<<blockpergrid, threadperblock>>>(d_input, d_output, N);
     hipDeviceSynchronize();
 
-    hipMemcpy(output, d_output, sizeof(float), hipMemcpyDeviceToHost);
+    hipMemcpy(output, d_output, sizeof(int), hipMemcpyDeviceToHost);
 
     hipFree(d_input);
     hipFree(d_output);
@@ -48,10 +48,10 @@ int main(int argc, char* argv[]) {
         return 1;
     }
     int N;
-    float output;
+    int output;
     
     input_file >> N;
-    std::vector<float> input(N);
+    std::vector<int> input(N);
 
     for(int i = 0; i < N; ++i) {
         input_file >> input[i];
