+++ meduim/reduction/fs_optimize_sharedmem.cu
@@ -7,8 +7,8 @@
 
 #define threadperblock 256
 
-__global__ void reduction(const float* input, float* output, int N){
-    __shared__ float sdata[threadperblock];
+__global__ void reduction(const int* input, int* output, int N){
+    __shared__ int sdata[threadperblock];
 
     int idx = blockDim.x * blockIdx.x + threadIdx.x;
     int tidx = threadIdx.x;
@@ -27,23 +27,23 @@ __global__ void reduction(const float* input, float* output, int N){
 }
 
 
-extern "C" void solve(const float* input, float* output, int N) {  
+extern "C" void solve(const int* input, int* output, int N) {  
 
-    float *d_input, *d_output;
+    int *d_input, *d_output;
 
-    hipMalloc(&d_input, N * sizeof(float));
-    hipMalloc(&d_output, sizeof(float));
+    hipMalloc(&d_input, N * sizeof(int));
+    hipMalloc(&d_output, sizeof(int));
 
-    hipMemcpy(d_input, input, N * sizeof(float), hipMemcpyHostToDevice);
-    hipMemset(d_output, 0, sizeof(float));
+    hipMemcpy(d_input, input, N * sizeof(int), hipMemcpyHostToDevice);
+    hipMemset(d_output, 0, sizeof(int));
     
     int blockpergrid = (N + threadperblock - 1) / threadperblock;
 
-    //size_t shared_mem_size = threadperblock * sizeof(float);
+    //size_t shared_mem_size = threadperblock * sizeof(int);
     reduction<<<blockpergrid, threadperblock>>>(d_input, d_output, N);
     hipDeviceSynchronize();
 
-    hipMemcpy(output, d_output, sizeof(float), hipMemcpyDeviceToHost);
+    hipMemcpy(output, d_output, sizeof(int), hipMemcpyDeviceToHost);
 
     hipFree(d_input);
     hipFree(d_output);
@@ -64,10 +64,10 @@ int main(int argc, char* argv[]) {
         return 1;
     }
     int N;
-    float output;
+    int output;
     
     input_file >> N;
-    std::vector<float> input(N);
+    std::vector<int> input(N);
 
     for(int i = 0; i < N; ++i) {
         input_file >> input[i];
